+++ CHANGELOG.md
@@ -24,6 +24,11 @@
 
 
 ### Deprecated
+
+* **scrollback** option in `footrc`. Use the **lines** option in the
+  **[scrollback]** section instead.
+
+
 ### Removed
 ### Changed